fix: syncloop issues
Currently `sync.startOnce()` is never cancelled, even though the
syncLoopItem is not in the queue anymore, this commit fixes this.

Also, previous pattern was to stop the syncLoop and then start it again
when the parameters of `sync.start()` should change.

Now changing parameters can just be done by calling
`sync.start(newparams)`, the sync loop will restart with new parameters
similar to how syncOnce cancels the current loop and syncs once with the
specified settings.